requirements.txt
streamlit
pandas
boto3
Pillow
